requirements.txt
pytest==8.3.2
selenium
webdriver-manager
allure-pytest
